requirements.txt
py-cord==2.9.0
Pillow==10.4.0
pytimeparse==1.1.8
matplotlib==3.9.2
